+++ example/routes/auth.js
@@ -1,34 +1,25 @@
-const express = require('express');
-const router = express.Router();
+const router = require('express').Router();
 const {BearerParser, BearerValidator} = require('../../dist/build.common');
 // const {BearerParser, BearerValidator} = require('bearer-token-parser');
 
-/**
- * Token authentication.
- */
 router.get('/', [
-  // Validate Bearer tokens.
+  // Token authentication middleware initialization.
   BearerValidator.validation({
-    realm: 'Sample API',
+    realm: 'myapi',
     tokenCheckCallback: token => {
-      // Returns TRUE if the token is correct.
-      return token === 'eTRPXY8F~np0zbAzi2~KN';
+      // Return `TRUE` if the token is correct.
+      // If you return `FALSE`, a `401` error will be returned by the `BearerValidator.validation` middleware.
+      return token === 'mytoken123';
     }
   })
 ], (req, res) => {
   // Get bearer token.
-  // It gets a token68 format token from the authorization header of the request.
   const token = BearerParser.parseBearerToken(req.headers);
-  console.log(`token=${token}`);
   res.send('Authentication was successful');
 });
 
-/**
- * Get tokens without authenticating.
- */
- router.get('/parse', (req, res) => {
+router.get('/parse', (req, res) => {
   const token = BearerParser.parseBearerToken(req.headers);
-  console.log(`token=${token}`);
   res.send(`The token obtained is ${token||'undefined'}`);
 });