+++ src/BearerValidator.ts
@@ -1,122 +1,103 @@
+import express from "express";
 import BearerParser from '~/BearerParser';
-import {Request, Response, NextFunction} from "express";
+import BearerValidatorOptions from '~/interfaces/BearerValidatorOptions';
+import isAsyncFunction from '~/utils/isAsyncFunction';
 
 /**
- * This is middleware that validates Bearer token authentication with the Express framework.
- *
- * @example
- * import express from 'express';
- * import {body, validationResult} from 'express-validator';
- * import {BearerParser, BearerTokenValidator} from 'bearer-token-parser';
- * 
- * const router = express.Router();
- * router.post('/', [
- *   // Validate input data.
- *   body('email').isEmail(),
- *   body('name').isLength({min: 1, max: 20}),
- * 
- *   // Validate Bearer tokens.
- *   BearerTokenValidator.validation({
- *     realm: 'Sample API',
- *     tokenCheckCallback: async (token) => {
- *       // Returns TRUE if the token is correct.
- *       return token === '<Your Bearer token>';
- *     },
- *     requestParameterCheck: (req) => {
- *       // Returns TRUE if the input data is correct using the "express-validator" package..
- *       const errors = validationResult(req);
- *       return errors.isEmpty();
- *     }
- *   }),
- * ], async (req, res, next) => {
- * 
- *   // Get bearer token.
- *   const token = BearerParser.parseBearerToken(req.headers);
- *   console.log(`Token: ${token}`);
- * 
- *   // Processing something.
- *   // ...
- * 
- *   // Respond.
- *   res.json(true);
- * });
- * 
- * // mount the router on the app
- * app.use('/', router)
+ * Express framework middleware for Bearer token authentication.
  */
 export default class {
-
   /**
-   * Returns a middleware function that checks the bearer token.
-   * 
-   * @param  {Object} opts
-   * @return {Function}
+   * Token authentication middleware initialization.
+   * @param {BearerValidatorOptions} options Bearer token validation option.
+   * @return {(req: express.Request, res: express.Response, next: express.NextFunction) => Promise<void>} Token authentication middleware.
+   * @throws {@link TypeError} If the token location is `query` or `body` and no token parameter is specified.
+   * @throws {@link TypeError} Invalid value other than `header`, `query`, or `body` specified for token location.
    */
-  public static validation(opts: {
-    realm?: string,
-    tokenCheckCallback?: undefined|Function,
-    requestParameterCheck?: undefined|Function
-  }): Function {
-
+  public static validation(options: BearerValidatorOptions): (req: express.Request, res: express.Response, next: express.NextFunction) => Promise<void> {
     // Initialize options.
-    opts = Object.assign({
+    options = Object.assign({
+      tokenLocation: 'header',
+      tokenParameter: 'access_token',
       realm: '',
       tokenCheckCallback: undefined,
       requestParameterCheck: undefined
-    }, opts || {});
+    }, options || {});
 
-    // Returns a middleware function that checks the bearer token.
-    return async (req: Request, res: Response, next: NextFunction) => {
-      // Returns a 401 error if the request does not include an Authorization header.
-      if (!req.headers.authorization)
-          return void res
-            .header('WWW-Authenticate', `Bearer realm="${opts.realm}", error="token_required"`)
-            // .header('WWW-Authenticate', 'Bearer realm="token_required"')
-            .sendStatus(401);
+    // Required check for token parameter name.
+    if ((options.tokenLocation === 'query' || options.tokenLocation === 'body') && !options.tokenParameter)
+      throw new TypeError('If the token location is `query` or body`, the token parameter name is required');
 
-      // Get bearer token.
-      const token = BearerParser.parseBearerToken(req.headers);
+    // Returns middleware that validates bearer tokens.
+    return async (req: express.Request, res: express.Response, next: express.NextFunction) => {
+      // If the token location is `Header`, return a `401` error if the `Authorization` header is not present.
+      // If the token location is `Body`/`Query`, return a `401 error if the token parameter is not present.
+      if (
+        (options.tokenLocation === 'header' && (!req.headers || !req.headers.authorization))
+        || (options.tokenLocation === 'query' && (!req.query || !req.query.hasOwnProperty(options.tokenParameter!)))
+        || (options.tokenLocation === 'body' && (!req.body || !req.body.hasOwnProperty(options.tokenParameter!)))
+      )
+        return void this.resWithWwwAuthenticate(res, 401, options.realm!, 'token_required');
 
-      // Returns a 401 error if the token in the Authorization header is empty or malformatted.
+      // Get Token.
+      let token;
+      switch (options.tokenLocation) {
+      case 'header':
+        token = BearerParser.parseBearerTokenHeader(req);
+        break;
+      case 'query':
+        token = BearerParser.parseBearerTokenQuery(req, options.tokenParameter);
+        break;
+      case 'body':
+        token = BearerParser.parseBearerTokenBody(req, options.tokenParameter);
+        break;
+      default:
+        throw new TypeError('token location is invalid value, can be any of `header`, `query`, or `body`');
+      }
+
       if (!token)
-          return void res
-            .header('WWW-Authenticate', `Bearer realm="${opts.realm}", error="invalid_token", error_description="Token format error"`)
-            .sendStatus(401);
-
-      // Check if the token is correct.
-      if (opts.tokenCheckCallback) {
-        const valid = this.isAsyncFunction(opts.tokenCheckCallback)
-          ? await opts.tokenCheckCallback(token)
-          : opts.tokenCheckCallback(token);
+        // If the token is empty or the format is invalid, a `401` error is returned.
+        return void this.resWithWwwAuthenticate(res, 401, options.realm!, 'invalid_token', 'Token format error');
+      if (options.tokenCheckCallback) {
+        // Execute token validation callback function.
+        const valid = isAsyncFunction(options.tokenCheckCallback) ? await options.tokenCheckCallback(token) : options.tokenCheckCallback(token);
         if (!valid)
-            return void res
-              .header('WWW-Authenticate', `Bearer realm="${opts.realm}", error="invalid_token", error_description="Token cannot be authenticated"`)
-              .sendStatus(401);
+          return void this.resWithWwwAuthenticate(res, 401, options.realm!, 'invalid_token', 'Token cannot be authenticated');
+      }
+      if (options.requestParameterCheck) {
+        // Execute request data validation callback function.
+        const valid = isAsyncFunction(options.requestParameterCheck) ? await options.requestParameterCheck(req) : options.requestParameterCheck(req);
+        if (!valid)
+          return void this.resWithWwwAuthenticate(res, 400, options.realm!, 'invalid_request');
       }
 
-      // Check request data.
-      if (opts.requestParameterCheck) {
-        const valid = this.isAsyncFunction(opts.requestParameterCheck)
-          ? await opts.requestParameterCheck(req)
-          : opts.requestParameterCheck(req);
-        if (!valid)
-          return void res
-            .header('WWW-Authenticate', `Bearer realm="${opts.realm}", error="invalid_request"`)
-            .sendStatus(400);
-      }
-
-      // If you can get a bearer token, proceed to the next middleware.
+      // If the authentication is successful, control is passed to the next process.
       next();
     };
   }
 
   /**
-   * Returns whether it is an Async function
-   * 
-   * @param  {Function} value
-   * @return {boolean}
+   * HTTP response with `WWW-Authenticate` header.
+   * @param {express.Response} res Response object.
+   * @param {number} status HTTP Status Code.
+   * @param {string} realm Realm name to be included in response headers.
+   * @param {string|undefined} error? Error Name. Set to `error` in the `WWW-Authenticate` header. The default is none (`undefined`).
+   * @param {string|undefined} errorDescription? Error Description. Set to `error_description` in the `WWW-Authenticate` header. The default is none (`undefined`).
    */
-  private static isAsyncFunction(value: Function): boolean {
-    return value && value.constructor && value.constructor === Object.getPrototypeOf(async function(){}).constructor
+  public static resWithWwwAuthenticate(
+    res: express.Response,
+    status: number,
+    realm: string,
+    error?: string,
+    errorDescription?: string
+  ) {
+    let wwwAuthenticate = `Bearer realm="${realm}"`;
+    if (error)
+      wwwAuthenticate += `, error="${error}"`;
+    if (errorDescription)
+      wwwAuthenticate += `, error_description="${errorDescription}"`;
+    res
+      .header('WWW-Authenticate', wwwAuthenticate)
+      .sendStatus(status);
   }
 }