+++ src/index.ts
@@ -1,9 +1,4 @@
-import { version } from '~/package';
 import BearerParser from '~/BearerParser';
 import BearerValidator from '~/BearerValidator';
 
-export {
-  version,
-  BearerParser,
-  BearerValidator
-}
+export {BearerParser, BearerValidator}