+++ types/BearerValidator.d.ts
@@ -1,63 +1,24 @@
+import express from "express";
+import BearerValidatorOptions from '~/interfaces/BearerValidatorOptions';
 /**
- * This is middleware that validates Bearer token authentication with the Express framework.
- *
- * @example
- * import express from 'express';
- * import {body, validationResult} from 'express-validator';
- * import {BearerParser, BearerTokenValidator} from 'bearer-token-parser';
- *
- * const router = express.Router();
- * router.post('/', [
- *   // Validate input data.
- *   body('email').isEmail(),
- *   body('name').isLength({min: 1, max: 20}),
- *
- *   // Validate Bearer tokens.
- *   BearerTokenValidator.validation({
- *     realm: 'Sample API',
- *     tokenCheckCallback: async (token) => {
- *       // Returns TRUE if the token is correct.
- *       return token === '<Your Bearer token>';
- *     },
- *     requestParameterCheck: (req) => {
- *       // Returns TRUE if the input data is correct using the "express-validator" package..
- *       const errors = validationResult(req);
- *       return errors.isEmpty();
- *     }
- *   }),
- * ], async (req, res, next) => {
- *
- *   // Get bearer token.
- *   const token = BearerParser.parseBearerToken(req.headers);
- *   console.log(`Token: ${token}`);
- *
- *   // Processing something.
- *   // ...
- *
- *   // Respond.
- *   res.json(true);
- * });
- *
- * // mount the router on the app
- * app.use('/', router)
+ * Express framework middleware for Bearer token authentication.
  */
 export default class {
     /**
-     * Returns a middleware function that checks the bearer token.
-     *
-     * @param  {Object} opts
-     * @return {Function}
+     * Token authentication middleware initialization.
+     * @param {BearerValidatorOptions} options Bearer token validation option.
+     * @return {(req: express.Request, res: express.Response, next: express.NextFunction) => Promise<void>} Token authentication middleware.
+     * @throws {@link TypeError} If the token location is `query` or `body` and no token parameter is specified.
+     * @throws {@link TypeError} Invalid value other than `header`, `query`, or `body` specified for token location.
      */
-    static validation(opts: {
-        realm?: string;
-        tokenCheckCallback?: undefined | Function;
-        requestParameterCheck?: undefined | Function;
-    }): Function;
+    static validation(options: BearerValidatorOptions): (req: express.Request, res: express.Response, next: express.NextFunction) => Promise<void>;
     /**
-     * Returns whether it is an Async function
-     *
-     * @param  {Function} value
-     * @return {boolean}
+     * HTTP response with `WWW-Authenticate` header.
+     * @param {express.Response} res Response object.
+     * @param {number} status HTTP Status Code.
+     * @param {string} realm Realm name to be included in response headers.
+     * @param {string|undefined} error? Error Name. Set to `error` in the `WWW-Authenticate` header. The default is none (`undefined`).
+     * @param {string|undefined} errorDescription? Error Description. Set to `error_description` in the `WWW-Authenticate` header. The default is none (`undefined`).
      */
-    private static isAsyncFunction;
+    static resWithWwwAuthenticate(res: express.Response, status: number, realm: string, error?: string, errorDescription?: string): void;
 }