+++ types/index.d.ts
@@ -1,4 +1,3 @@
-import { version } from '~/package';
 import BearerParser from '~/BearerParser';
 import BearerValidator from '~/BearerValidator';
-export { version, BearerParser, BearerValidator };
+export { BearerParser, BearerValidator };