+++ types/utils/isAsyncFunction.d.ts
@@ -1,0 +1,7 @@
+/**
+ * Returns whether it is an Async function
+ * @param {Function} payload Function.
+ * @return {boolean} `TRUE` for asynchronous functions, `FALSE` otherwise.
+ */
+declare const _default: (payload: Function) => boolean;
+export default _default;